Create Angular package exposing ApiEditorComponent
This adds `public_api.ts` file that exports the `ApiEditorComponent`
along with all the dependencies. To create the package invoke
`yarn package` that uses ng-packagr to help with the ahead of time
compilation and bundling. The resulting Angular package is created in
the root directory of the project as `dist.tgz`.