Implement channel name regex configuration
Hi,
this PR allows externalizing the channel formalisation regex found in add-channel.component.html.
The regex will be in the DEFAULT_CONFIG of the config.service.ts file and editable with the VALIDATION_CHANNELNAME_REGEXP_ENV environment variable.

However, it must be taken into account that this regex will cross the different build layers.